Refactor the super components
