+++ hayhooks/tests/components/test_search_extraction.py
@@ -1,14 +1,13 @@
 from haystack import Document
 
-from components.content_extraction import build_search_extraction_component
+from components.web_search.document_extract import DocumentContentExtractor
 
 
 def test_search_extraction_component():
     """Test search extraction content"""
 
-    component = build_search_extraction_component()
+    component = DocumentContentExtractor()
     doc = Document.from_dict({"title": "title", "content": "derp", "link": "http://example.com"})
-    kwargs = {"documents": [doc]}
-    result = component.run(**kwargs)
+    result = component.run(documents=[doc])
     documents = result["documents"]
     assert documents[0].content.startswith("This domain is for use")
